Enum actix_wamp::ErrorKind

pub enum ErrorKind {
    InvalidURI,
    NoSuchProcedure,
    ProcedureAlreadyExists,
    NoSuchRegistration,
    NoSuchSubscription,
    InvalidArgument,
    SystemShutdown,
    CloseRealm,
    GoodbyeAndOut,
    NotAuthorized,
    AuthorizationFailed,
    NoSuchRealm,
    NoSuchRole,
    Canceled,
    OptionNotAllowed,
    NoEligibleCallee,
    NetworkFailure,
    Other(String),
}

WAMP pre-defined error URIs.

WAMP peers MUST use only the defined error messages.

Variants
InvalidURI

Peer provided an incorrect URI for any URI-based attribute of WAMP message, such as realm, topic or procedure

WAMP uri: wamp.error.invalid_uri

NoSuchProcedure

A Dealer could not perform a call, since no procedure is currently registered under the given URI.

WAMP uri: wamp.error.no_such_procedure

ProcedureAlreadyExists

A procedure could not be registered, since a procedure with the given URI is already registered.

WAMP uri: wamp.error.procedure_already_exists

NoSuchRegistration

A Dealer could not perform an unregister, since the given registration is not active.

WAMP uri: wamp.error.no_such_registration

NoSuchSubscription

A Broker could not perform an unsubscribe, since the given subscription is not active.

WAMP uri: wamp.error.no_such_subscription

InvalidArgument

A call failed since the given argument types or values are not acceptable to the called procedure. In this case the Callee may throw this error. Alternatively a Router may throw this error if it performed payload validation of a call, call result, call error or publish, and the payload did not conform to the requirements.

WAMP uri: wamp.error.invalid_argument

SystemShutdown

The Peer is shutting down completely - used as a GOODBYE (or ABORT) reason.

WAMP uri: wamp.error.system_shutdown

CloseRealm

The Peer want to leave the realm - used as a GOODBYE reason.

WAMP uri: wamp.error.close_realm

GoodbyeAndOut

A Peer acknowledges ending of a session - used as a GOODBYE reply reason.

WAMP uri: wamp.error.goodbye_and_out

NotAuthorized

A join, call, register, publish or subscribe failed, since the Peer is not authorized to perform the operation.

WAMP uri: wamp.error.not_authorized

AuthorizationFailed

A Dealer or Broker could not determine if the Peer is authorized to perform a join, call, register, publish or subscribe, since the authorization operation itself failed. E.g. a custom authorizer did run into an error.

WAMP uri: wamp.error.authorization_failed

NoSuchRealm

Peer wanted to join a non-existing realm (and the Router did not allow to auto-create the realm).

WAMP uri: wamp.error.no_such_realm

NoSuchRole

A Peer was to be authenticated under a Role that does not (or no longer) exists on the Router. For example, the Peer was successfully authenticated, but the Role configured does not exists - hence there is some misconfiguration in the Router.

WAMP uri: wamp.error.no_such_role

Canceled

uriDealer or Callee canceled a call previously issued

WAMP uri: wamp.error.canceled

OptionNotAllowed

A Peer requested an interaction with an option that was disallowed by the Router

WAMP uri: wamp.error.option_not_allowed

NoEligibleCallee

A Dealer could not perform a call, since a procedure with the given URI is registered, but Callee Black- and Whitelisting and/or Caller Exclusion lead to the exclusion of (any) Callee providing the procedure.

WAMP uri: wamp.error.no_eligible_callee

NetworkFailure

A Router encountered a network failure

WAMP uri: wamp.error.network_failure

Other(String)

Any other error code.

Methods

impl ErrorKind

pub fn uri(&self) -> &str

pub fn from_uri(uri: &str) -> Self